chore: migrate storybook config to ts
