+++ package.json
@@ -118,7 +118,6 @@
     "@size-limit/preset-big-lib": "7.0.5",
     "@storybook/addon-a11y": "6.4.13",
     "@storybook/addon-actions": "6.4.13",
-    "@storybook/addon-docs": "6.4.13",
     "@storybook/addon-essentials": "6.4.13",
     "@storybook/addon-links": "6.4.13",
     "@storybook/addons": "6.4.13",
